:recycle: move transactionDetailView to it's own folder
