+++ src/components/transactions/transactionDetailView/transactionDetailView.test.js
@@ -2,9 +2,9 @@ import React from 'react';
 import thunk from 'redux-thunk';
 import { expect } from 'chai';
 import { spy } from 'sinon';
-import { mountWithContext } from '../../../test/utils/mountHelpers';
+import { mountWithContext } from '../../../../test/utils/mountHelpers';
 import TransactionDetailView from './transactionDetailView';
-import txTypes from '../../constants/transactionTypes';
+import txTypes from '../../../constants/transactionTypes';
 
 describe('TransactionDetailView', () => {
   it('should render 5 rows', () => {
