+++ src/components/transactions/transactionDetailView/transactionDetailViewRow.js
